--- a/src/python/pants/backend/project_info/rules/dependencies.py
+++ b/src/python/pants/backend/project_info/rules/dependencies.py
@@ -61,7 +61,9 @@ async def dependencies(
     else:
         hydrated_targets = await Get[HydratedTargets](Addresses, addresses)
         address_strings.update(
-            dep.spec for hydrated_target in hydrated_targets for dep in hydrated_target.dependencies
+            dep.spec
+            for hydrated_target in hydrated_targets
+            for dep in hydrated_target.adaptor.dependencies
         )
 
     with options.line_oriented(console) as print_stdout: