--- a/src/python/pants/backend/project_info/rules/dependencies.py
+++ b/src/python/pants/backend/project_info/rules/dependencies.py
@@ -56,7 +56,7 @@ async def dependencies(
         transitive_targets = await Get[TransitiveHydratedTargets](Addresses, addresses)
         transitive_dependencies = transitive_targets.closure - set(transitive_targets.roots)
         address_strings.update(
-            hydrated_target.address.spec for hydrated_target in transitive_dependencies
+            hydrated_target.adaptor.address.spec for hydrated_target in transitive_dependencies
         )
     else:
         hydrated_targets = await Get[HydratedTargets](Addresses, addresses)