--- a/src/python/pants/engine/legacy/BUILD
+++ b/src/python/pants/engine/legacy/BUILD
@@ -88,11 +88,3 @@ python_library(
   ],
   tags = {"partially_type_checked"},
 )
-
-python_tests(
-  name='tests',
-  dependencies=[
-    ':graph'
-  ],
-  tags = {"partially_type_checked"},
-)