… src/python/pants/backend/python/lint/pylint/rules_integration_test.py
… src/python/pants/backend/python/rules/run_setup_py_test.py
… src/python/pants/engine/legacy/graph_test.py
… src/python/pants/rules/core/fmt_test.py
… src/python/pants/rules/core/list_targets.py
… src/python/pants/rules/core/test.py
… src/python/pants/rules/core/test_test.py