### Problem

`HydratedTarget` and `TargetAdaptor` currently both duplicate the exact same information in the fields `address` and `dependencies`, i.e. `hydrated_target.address == hydrated_target.adaptor.address` and `hydrated_target.dependencies == hydrated_target.adaptor.dependencies`.

Why keep both these types, then? We need `HydratedTarget` as a simple wrapper around `TargetAdaptor` due to the engine being a type-driven API. Sometimes, we want a generic way to refer to a target, regardless of what the target type is, i.e. `HydratedTarget`. Other times, we do care about the target type, such as a `PythonTestsAdaptor` vs. `PythonLibraryAdaptor`.

So, we need both types, but duplicating the information in both types makes the distinction between the two less clear—we want to emphasize that `HydratedTarget` is nothing more than a wrapper around `TargetAdaptor` for the purpose of having a generic target type for the engine. Duplicating the fields also complicates the implementation, e.g. making it harder to use `HydratedTarget` in tests.

### Solution

Simplify `HydratedTarget` to have only one public* field of `adaptor`. Whenever we need to access the address or dependencies, we can use `hydrated_target.adaptor.address` and `hydrated_target.adaptor.dependencies`.

*`HydratedTarget` gets passed around hundreds of times to the engine when constructing the build graph, so it needs to be hashed hundreds of times. To improve the speed of this, we override `HydratedTarget.__hash__` to only consider the address. However, changing `__hash__` to use `adaptor.address` instead of a direct field `.address` results in a several-second performance hit when running `build-support/bin/mypy.py` (which runs on every target in Pants). Why? `TargetAdaptor.address` is ~10x slower than `HydratedTarget.address` (0.000_000_1 vs. 0.000_001 seconds). So, we add a private field `HydratedTarget._address` to avoid this performance regression when hashing. Even better, this PR improves the time to build the build graph with `build-support/bin/mypy.py` from 4.7 seconds to 3.8 seconds by implementing `__eq__`, which should have been implemented from the start.

### Result

The difference between `HydratedTarget` and `TargetAdaptor` is hopefully more explicit.

The time to build the build graph with `build-support/bin/mypy.py` improved from 4.7 seconds to 3.8 seconds.

This is sort of prework for https://github.com/pantsbuild/pants/issues/4535 as it simplifies the status quo.